some standardization of load_lmpdat
