+++ tests/test_atoms.py
@@ -115,95 +115,6 @@ def test_atoms_extend__reindexes_new_impropers_to_proper_atoms(linear_cnnc):
     double_cnnc.extend(linear_cnnc)
     assert (double_cnnc.impropers == [(0,1,2,3), (4,5,6,7)]).all()
 
-def test_atoms_save_lmpdat__from_load_cif_is_successful():
-    with importlib.resources.path(tests, "uio66.cif") as path:
-        uio66 = Atoms.load_cif(path)
-        uio66.atom_type_labels = uio66.atom_type_elements
-
-    sout = io.StringIO("")
-    uio66.save_lmpdat(sout)
-
-def test_atoms_load_lmpdat__uio66_has_arrays_of_right_size():
-    with importlib.resources.open_text(tests, "uio66-F.lmpdat") as f:
-        atoms = Atoms.load_lmpdat(f, atom_format="full", use_comment_for_type_labels=True)
-
-    assert len(atoms.atom_type_masses) == 4
-    assert len(atoms.pair_coeffs) == 4
-    assert len(atoms.bond_type_coeffs) == 2
-    assert len(atoms.angle_type_coeffs) == 2
-    assert atoms.positions.shape == (16,3)
-    assert len(atoms.groups) == 16
-    assert len(atoms.charges) == 16
-    assert len(atoms.atom_types) == 16
-    assert len(atoms.bonds) == 4
-    assert len(atoms.angles) == 8
-
-
-def test_atoms_save_lmpdat__with_no_atom_type_labels_outputs_file_with_type_id_labels():
-    with importlib.resources.open_text(tests, "uio66-linker-arb-terms-no-labels.lmpdat") as f:
-        atoms = Atoms.load_lmpdat(f, atom_format="full", use_ids_for_type_labels_and_elements=True)
-
-    sout = io.StringIO("")
-    atoms.save_lmpdat(sout, file_comment="uio66-linker-arb-terms-no-labels.lmpdat")
-
-    # output file code, in case we need to update the lmpdat file because of new format changes
-    # with open("uio66-linker.lmpdat", "w") as f:
-    #     sout.seek(0)
-    #     f.write(sout.read())
-
-    with importlib.resources.open_text(tests, "uio66-linker-arb-terms-atom-type-id-labels.lmpdat") as f:
-        sout.seek(0)
-        assert sout.read() == f.read()
-
-def test_atoms_load_lmpdat__no_atom_type_labels():
-    with importlib.resources.open_text(tests, "uio66-linker-arb-terms-no-labels.lmpdat") as f:
-        atoms = Atoms.load_lmpdat(f, atom_format="full", use_ids_for_type_labels_and_elements=True)
-
-
-def test_atoms_load_lmpdat__use_comment_for_type_labels_with_no_atom_type_labels_raises_exception():
-    with importlib.resources.open_text(tests, "uio66-linker-arb-terms-no-labels.lmpdat") as f:
-        with pytest.raises(Exception):
-            atoms = Atoms.load_lmpdat(f, atom_format="full", use_comment_for_type_labels=True)
-
-def test_atoms_save_lmpdat__output_file_identical_to_one_read():
-    with importlib.resources.open_text(tests, "uio66-linker-arb-terms.lmpdat") as f:
-        sin = StringIO(f.read())
-        uio66_linker_ld = Atoms.load_lmpdat(sin, atom_format="full", use_comment_for_type_labels=True)
-
-    sout = io.StringIO("")
-    uio66_linker_ld.save_lmpdat(sout, file_comment="uio66-linker-arb-terms.lmpdat")
-
-    # output file code, in case we need to update the lmpdat file because of new format changes
-    # with open("uio66-linker-arb-terms.lmpdat", "w") as f:
-    #     sout.seek(0)
-    #     f.write(sout.read())
-
-    sout.seek(0)
-    sin.seek(0)
-    assert sout.read() == sin.read()
-
-def test_atoms_load__loads_lmpdat_from_file_or_path():
-    with importlib.resources.path(tests, "uio66-linker.lmpdat") as path:
-        atoms = Atoms.load(path, atom_format="atomic")
-        assert len(atoms) == 16
-
-    with importlib.resources.open_text(tests, "uio66-linker.lmpdat") as fd:
-        atoms = Atoms.load(fd, filetype="lmpdat", atom_format="atomic")
-        assert len(atoms) == 16
-
-def test_atoms_load__loads_cml():
-    with importlib.resources.path(tests, "uio66-linker.cml") as path:
-        atoms = Atoms.load(path)
-        assert len(atoms) == 16
-
-def test_atoms_load__loads_cif_from_file_or_path():
-    with importlib.resources.path(tests, "uio66-linker.cif") as path:
-        atoms = Atoms.load(path)
-        assert len(atoms) == 16
-
-    with importlib.resources.open_text(tests, "uio66-linker.cif") as fd:
-        atoms = Atoms.load(fd, filetype="cif")
-        assert len(atoms) == 16
 
 def test_find_unchanged_atom_pairs__same_structure_is_unchanged(linear_cnnc):
     assert find_unchanged_atom_pairs(linear_cnnc, linear_cnnc) == [(0,0), (1,1), (2,2), (3,3)]
@@ -227,20 +138,6 @@ def test_atoms_getitem__has_all_atom_types_and_charges():
     assert atoms[(0,1)].elements == ["N", "C"]
     assert atoms[(2,3)].elements == ["C", "N"]
     assert (atoms[(0,1)].charges ==[1, 2]).all()
-
-def test_atoms_load_cml__loads_elements_bonds():
-    with importlib.resources.path(tests, "uio66-linker.cml") as path:
-        atoms = Atoms.load_cml(path)
-
-    assert atoms.elements == ["C", "O", "O", "C", "C", "H", "H", "C", "C", "C", "C", "H", "H", "O", "C", "O"]
-    assert (atoms.bonds == [[0, 1], [10, 11], [0, 2], [0, 3], [3, 10], [9, 10], [3, 4], [9, 12],
-                           [8, 9], [4, 5], [4, 7], [7, 8], [8, 14], [6, 7], [13, 14], [14, 15]]).all()
-
-def test_atoms_load_cif__loads_elements():
-    with importlib.resources.open_text(tests, "uio66-linker.cif") as fd:
-        atoms = Atoms.load_cif(fd)
-
-    assert atoms.elements == ["C", "O", "O", "C", "C", "H", "H", "C", "C", "C", "C", "H", "H", "O", "C", "O"]
 
 def test_atoms_calc_angles__ethane_has_12_angles():
     ethane = Atoms(elements='HHHCCHHH', positions=[(1., 1., 1)] * 8,
