--- tests/uio66-F.lmpdat
@@ -1,72 +1,0 @@
- (written by mofun)
-
-16 atoms
-4 bonds
-8 angles
-0 dihedrals
-0 impropers
-
-4 atom types
-2 bond types
-2 angle types
-
-Masses
-
- 1 12.0107   # C_2
- 2 12.0107   # C_R
- 3 15.9994   # O_1
- 4 18.9984   # F_
-
-Pair Coeffs
-
- 1   0.105000   3.430851 # C_2
- 2   0.105000   3.430851 # C_R
- 3   0.060000   3.118146 # O_1
- 4   0.050000   2.996983 # F_
-
-Bond Coeffs
-
- 1 443.203299   1.354566 # C_R F_
- 2 443.203299   1.354566 # F_ C_R
-
-Angle Coeffs
-
- 1 cosine/periodic 207.488907 -1 3 # F_ C_R C_R
- 2 cosine/periodic 207.488907 -1 3 # C_R C_R F_
-
-Atoms
-
- 1 1 1   0.000000   0.000000   0.000000   0.000000   # C_2
- 2 1 3   0.000000   0.349000   0.000000  -1.207000   # O_1
- 3 1 3   0.000000  -1.207000   0.000000   0.349000   # O_1
- 4 1 2   0.000000   1.064000   0.000000   1.064000   # C_R
- 5 1 2   0.000000   0.637000   0.000000   2.358000   # C_R
- 6 2 4   0.000000  -0.479000   0.000000   2.590000   # F_
- 7 2 4   0.000000   1.401000   0.000000   4.469000   # F_
- 8 1 2   0.000000   1.633000   0.000000   3.353000   # C_R
- 9 1 2   0.000000   2.927000   0.000000   2.927000   # C_R
- 10 1 2   0.000000   3.353000   0.000000   1.633000   # C_R
- 11 1 2   0.000000   2.358000   0.000000   0.637000   # C_R
- 12 2 4   0.000000   2.590000   0.000000  -0.479000   # F_
- 13 2 4   0.000000   4.469000   0.000000   1.401000   # F_
- 14 1 3   0.000000   5.198000   0.000000   3.642000   # O_1
- 15 1 1   0.000000   3.991000   0.000000   3.991000   # C_2
- 16 1 3   0.000000   3.642000   0.000000   5.198000   # O_1
-
-Bonds
-
- 1 1 11 12   # C_R F_
- 2 1 10 13   # C_R F_
- 3 1 5 6   # C_R F_
- 4 2 7 8   # F_ C_R
-
-Angles
-
- 1 1 12 11 4   # F_ C_R C_R
- 2 1 12 11 10   # F_ C_R C_R
- 3 2 11 10 13   # C_R C_R F_
- 4 1 13 10 9   # F_ C_R C_R
- 5 2 4 5 6   # C_R C_R F_
- 6 1 6 5 8   # F_ C_R C_R
- 7 2 5 8 7   # C_R C_R F_
- 8 2 9 8 7   # C_R C_R F_
